Added ALE
The concept of asynchronously linting and debugging my code as I write
it sounds intriguing.

I'm still playing around with `fzf-preview.vim`, as the default plugin
`fzf.vim` is easier to use. Some of the key mappings in
`fzf-preview.vim` are hard-coded, and I would prefer to change them so I
can bring them in line with my other mappings.